added support for running pylint with optional --jobs arg (#3)
This is a breaking change and the repo will be bumped to 2.0.0

* added support for running pylint with optional --jobs arg

* made pylint_test args more compatible with pylint

pylint_test now allows you to pass in multiple modules and also ignore multiple modules. Unfortunately this is a breaking change with what we had before as ignoring modules now (necessarily) requires use of the --ignore flag (as is the case with pylint). The reason for this change is that the --jobs flag, which can easily provide a factor of 2 speed up, requires that multiple modules be passed in so they can be linted concurrently

* updated README.md in anticpation of version bump